Refactored clocks into modules.
