+++ src/prelude.rs
@@ -1,5 +1,6 @@
 //! module to include background types or something idk
 
 pub use crate::clocks::ICSExt as _kea_hal_clocks_IcsExt;
+pub use crate::clocks::OSCExt as _kea_hal_clocks_OscExt;
 pub use crate::gpio::GPIOExt as _kea_hal_gpio_GpioExt;
 pub use crate::hal::prelude::*;
